[all] refactor resource handle
instead of hash-based resource handles, we want to switch to type-safe
opaque handles.

We do this so that we can create unnamed handles (which is useful for
helper modules, where creating name-hashed handles in a global namespace
would lead eventually to hash collisions which are really hard to debug.

Having opaque handles allows us to associate debug and metadata to the
pointer represented by the handle.