fixes for vulkan 1.3 (#40)
* fixes for vulkan 1.3

apply explicit casts where implicit conversion would not work anymore.

* apply missing changes - and reformat backend_vk

* [build] update cmake files

for static builds, we want to ensure that symbols are resolved
correctly. When using the gnu linker, and only then, the order of how
static libraries are passed to the linker is important as symbols will
only be resolved if they have been previously requested. we work around
this by setting the linker flags `--group-start` and `--group-end`
around the list of static libraries that we pass to the linker. This is
much easier than somehow coercing cmake to build the correct DAG for
liker dependencies.

Other linkers, most notably lld and mold, don't need these linker flags,
as they don't take the order or files on the linker line into account.

+ also cleaning up cmake files a bit
* update glm version
* [all] reformat using clang-format
* [ci] linux: add cmake prefix path
* [ci] update workflows to work with local vk_sdk