[backend] refactor vk extension require api
instead of passing required instance or device extensions to the backend
when the renderer gets initialised, we want to set these requirements
via a global singleton.

This makes it more easy for external modules to require vulkan
extensions, even without them having direct access to the renderer. As
long as required extensions get added to the global backend settings
object before the backend gets initialised, we are golden.

There can only ever be one backend. Therefore we can have settings for
the backend as a global singleton, which is accessible to all as long as
the backend has not yet been initialized.