[backend] split draw queue submission
split draw queue submission so that we have one submission which
dispatches the draw batch and signals semaphores on completion, and a
subsequent submission which waits on any timeline semaphores and signals
the fence on completion.

this is so that we may execute compute or transfer batches on parallel
queues which are synchronized via timeline semaphores. the same timeline
semaphores upon which we wait in the second submission to the draw
queue.

by signalling the fence after the second batch has finished waiting on
the timeline semaphores we can guarantee that any resources which are
protected by the frame fence are released as soon as the frame fence is
signalled.