[backend|renderer] refactor auto-sync
refactor how automatic sync works- we used to use LeRessourceUsageFlags
for resources, but this is not fine-grained enough - instead we want to
use le::AccessFlags2, and implicitly set these based on resource type,
so
that you only have to set them as an user of the API in special cases.

Things we found out during the refactor:

+ reads to image resources can imply an extra pre-access read/write
  operation in case the image first needs to perform an image layout
  transform. This means we must track any reads to images as potential
  read/write operations until we can somehow prove that there will be
  no layout transform.

This refactor is not yet complete, we still have some scaffolding in
there, and we also still accumulate usage flags over passes, which we
don't need to - usage flags should be set once with the rendergraph when
a resource is declared, and then be immutable. These flags need to cover
all possible accesses for the resource and cannot be extended once a
resource has been allocated.