[core] add logic to release callback forwarders
+ we want to be able to remove callback forwarders which are not used
anymore.

we use an occupancy bitfield which keeps track of which elements in the
table are in use.

before handing out a new table entry, we scan the occupancy bitfield for
the first available slot, which we set to true as soon as we hand out
the table entry.

when we release a table entry, we set the occupancy bit to false, which
means that it is available for next time when we hand out a new table
entry.