[backend] fix timeline semaphore signalling
timeline semaphores were not monotonically increasing their signalling
value, because i was (embarrassingly) takin a copy of the queue state
instead of updating the queue state in-place.

this is now fixed and mitigated by passing the queue state around
exclusively via pointers - this way you can't accidentally copy the
object.