[backend] enforce resources use same queue family
if we have multiple queue families active for an app, then we must make
sure that resources are not used by more than one queue family.

    realistically, this only affects read-only resources, and of these
    only read-only buffers.

we do this by first associating all queue requirements for a submission
with all its resources.

this means that per-resource we will have a superset of all the queue
requirements for all submissions which use this resource.

then we apply these per-resource queue requirements back to the
submission, which means that submissions will contain the superset of
their resouce's queue requirements.

effectively this means that if a resouce is used by two submissions with
different queue families, after this process both submissions will have
the same queue requirements (namely the superset of their initial queue
requirements), which will be answered by the same queue family, as the
requirements are the same.