[backend] increment acquire wait in separate step
when we describe the wait semaphore conditions for the acquire step, we
must not immediately increment the wait semaphore - as this would mean
that queues that wait for this queue would wait for the acquire step on
this queue to complete instead of the release step.

we now correctly increment acquire wait values in a separate step.