[scripts] create app ignore windows errors
on windows, when called via wsl, python shutil.copytree returns Errno 13 -
we ignore these errors as they don't seem to break the tree copy process.

Ultimately, this is a hack, but it seems more related to wsl and and python
and hard to pin down. It would be better to have a more robust method / tool
to create application templates...