[console] implement tentative commands
implement a few commands for console to see how these feel.

I think it might be better to have one console per connection in the
end, and not one console that is mirrored to all connections. Having a
bit of state associated with a connection could allow us to have
connections which are json based, in case we want to script via a
web-interface for example