[le_settings] add method to manipulate settings
add a method to set settings

- this is not yet very elegant - we would like some form of type-safety
which we get by copying (all!) settings into a local variable, then look
up if we can find a matching setting, and then see if we can get the
correct type for this setting.

there needs to be a better way to do this.