[le_backend | le_rendergraph] update dot settings
update settings which control generating .dot files: using integers
allows us to specify how many dot files to generate. this can be useful
when we want to inspect a series of frames and we want to make sure not
to accidentally miss a frame.

the previous functionality of generating a single .dot file is still
there - just request a single file