[backend|renderer] refactor backend_frames
since backend frames and swapchain images are conceptually different
(backend frames are all in-flight data for a frame whereas swapchain
images are images held by a swapchain) the renderer should only have to
worry about backend frames - how these map to swapchain images is the
backend's responsibility.

This is also so that we can have swapchains with different number of
backing images active at the same time.