[backend|renderer|swapchain] implement replace swapchain
implement replacing swapchain if we detect that acquire or present fail.
there seems to be a vexing issue right now where a swapchain image will
be shared between old and new swapchain, but then deleted with old
swapchain.

this could be a bug in wsi but it is more likely a bug in this code - we
need to track this down some more, but we might need insight into what's
happening in wsi land...