[renderer|backend] implement get default swp resource
disambiguate getting the default swapchain image resource and getting
image resource for an invalidated swapchain.

getting the default swapchain resource cannot be the same as getting the
swapchain resource for an invalid (==0) swapchain handle.

if a swapchain gets deleted, its handle will be NULL, and we should not
return a swapchain image for a swapchain whose handle is NULL, but we
should then return an invalid (=empty image handle).