[rendergraph] early-out on invalid image handles
if an invalid image handle is passed to add_color_attachment, then the
call to add_color_attachment returns early. The same happens for the
call to add_depth_stencil_attachment.

This is so that we can respond to swapchain images becoming invalid for
example: if a swapchain gets deleted, getting its image resource will
return NULL, and using this image resource as a color attachment does
now not add this color attachment, which is what we would expect.