[backend|renderer] refactor calc dataframes count
number of data frames should mirror number of smallest swapchain images
count for all active swapchains. ideally, we would set this
automatically on startup, but since the number of available swapchain
images only becomes queryable once we have created a surface via window,
we cannot set this automatically currently.

we need to find a better way to do this