[backend|renderer] implement deferred swapchain creation
in case you want to add swapchains dynamically, you must first setup the
renderer with a prototype setting for a future swapchain.

this is so that, before the backend is initialised, we can can query the
swapchain for extensions and capabilities that the swapchain requires
from the backend.

if you set the defer_create flag on the swapchain_settings, it will not
auto-create the swapchain on renderer setup, but it will only use the
settings to query backend requirements & extensions.