[le_core] allow const types for LE_SETTING
With this change, settings which are meant to be set only once can be
declared as const types. This is useful for global program settings such
as whether to use validation layers or not.

A const value setting can only be set once, at initialisation. If a
setting is defined in multiple compilation units, it will be set
uniquely when the initialisation first gets executed.