[backend] fix issue with non-continuous bindings
- in case a descriptor set has placeholder bindings, we must make sure
that the correct binding gets updated/set. there was a bug with the
previous implementation.

we used to set the binding directly via descriptor_set[binding] but as
there is no guarantee that bindings in descriptor_set have a binding
number that is identical with their index in the descriptor_set vector,
we must lookup the correct binding by iterating over elements in
descriptor_set, comparing against each binding number.

placeholder bindings may occur if the shader compiler optimises away
descriptor bindings which are not used.