add check for push constants
in case push constants are declared in shader, but unused in shader
code, the shader compiler might optimise them away. In this case,
SetPushConstantData would have failed with a quite opaque validation
error.

This can be quite irritating as it is often not straightforwardly clear
whether a shader makes use of Push Constants or not, and changing and
hot-reloading a shader might also result in Push Constants being
optimised away in shader code.

This fix adds a check for availability of push constants to a pipeline
layout - so that an island command to set push constants where the
shader has none will show an error message instead of potentially a
crash. The error message also gives a hint of what's potentially going
on.

An error message like this would normally be issued on every frame - to
prevent log spills, we add a new macro `LE_DEBUG_ONCE_IF_CHANGED` which
allows us to do things only once (per program reload per call site).