[backend|renderer] implement on_frame_clear_callback mechanism
add a mechanism so that users of the rendergraph can register callbacks
which they want to be triggered once the frame that the rendergraph
describes has completed. (i.e. once the backend clears this frame).

we can use this to manage lifetimes for objects which have objects
currently owned by a backend frame.

for the video decoder, we use an intrusive pointer mechanism for example
so that as long as the video encoder is in use in any frame that frame
has part-ownership of the video encoder and the encoder cannot be
destroyed.

instead, whenever either the frame or another part-owner of the video
decoder relinquishes ownership, we decrement a reference counter in the
video decoder (intrusive smart pointer pattern) until the reference
count has dropped to zero, upon which we can finally really destroy and
deallocate the video decoder, as we know that no more objects depend on
it.