[backend] implement auto-creation of yuv samplers
if you annotate a texture name with "__ycbcr__" in shader code, Island
will understand that this texture is a yuv texture. Instead of using a
default sampler for this texture it will generate a sampler using the
backend's ycbcr conversion sampler. image views generated while binding
this image will also automatically be created using this conversion
sampler info.