[cmake] add error and warning if missing submodule
if we're missing the imgui submodule immediately issue a warning that
tells us that this is the case instead of proceeding and writing into
the non-existant directory.

writing into the directory causes the directory to exist, which means
that subsequent attempts at updating and cloning the submodule will
fail. These changes mitigate this by printing out a relevant error
message and halting cmake generation at the earliest possible moment.