[le_pixels] report non-loading as warning
Set log level for loading error to warning so that this does not
automatically trigger a breakpoint interrupt; we do this because this
behaviour (warning instead of error) is less disruptive for when using
resource-hot-reloading using the resource manager.

Whoever consumes the le_pixels object needs to make sure that it is
valid after create(), anyway.