[pixels] add abstract interface for image decoder
add abstract interface for image decoder to le_pixels - we want this
interface to be implemented by all image format decoders (perhaps exr,
maybe ktx etc).

This will allow us to declare per-file-extension interface adapters
(backed by separately loaded modules) to the resourcemanager, who will
interface with the image loaders exclusively via the abstract interface.

the benefit of this is admittedly convoluted architecture is that we
don't have to declare a dependency from the resource manager to all the
different image decoder adapters as we can declare them at run-time.

-- scenario:

+ the app modul depends (and uses) the resource manager plugin
+ the resource manager plug-in comes with a set of common image decoders
that are based on stbimage
+ we want to give access to openexr, but only if this is needed (it
  can take time and effort to compile openexr) -- in this case we load
  the openexr decoder le_modul as a dependency to our app module.
+ at runtime, the app module queries the openexr adapter from the
  openexr module, and it introduces this adapter to the resource
  manager. the resource manager will now open .exr files via this
  adapter if it finds files with ending .openexr.