[le_core] only zero-out api on initial load
once api has been loaded, a reload should not zero out the old api
function pointers.

If we don't completely clear the api on reload, we can use whatever
persists in the api struct. This can be useful if you want to keep
singletons alive, or if you want to use abstract interfaces, where you
would store a pointer-to-interface in the api struct.