[renderer] implement map[buffer|image]
implement a method for transfer encoder to map memory directly.

this allows us to forego unnecessary extra allocations- especially when
decoding images:

1. map memory for image upload
2. read pixels into directly this memory

you must not access the mapped memory regions outside the current
transfer pass - but that's fine - renderpasses should be read-only when
it comes to the outside world, and they certainly are not allowed to
persistently store any temporary objects that they have acquired.