[le_path] fix and refactor arc calculations
arc calculation had a bug when calculating the arc rotation. This was
because the implementation notes use mathematical notation for matrixes,
and we forgot to apply a TRANSPOSE to the basis matrix.

we need the transpose because glm matrices are given as
glm::mat2(column_0,colum_1) -- note COLUMN and not ROW... but when
creating a coordinate basis you want to set the row, so that it
multiplies

m_00 , m_01    *   v_0   =    m_00 * v_0 + m_00 * v_1
m_10 , m_11        v_1        m_10 * v_0 + m_11 * v_1

this is because we assume column vectors...

the trick is to think: if my vector is a column, then what i initialise
the matrix with is a column, too...

+ additionally, refactor common maths out of both arc methods