[backend|renderer|swapchain|examples] refactor swapchain setting
make swapchain settings an extensible struct

we want to hide as much information as possible with swapchain settings,
and to consolidate the settings to only the files that affect them. This
change brings swapchain settings to each file they belong to ( windowed
swapchain settigns are declared and defined in le_swapchain_khr, imaged
swapchain settings are declared and defined in le_swapchain_img.

This gives us more flexibility - and it disentangles the headers a bit,
as renderer.h and rendererr.hpp don't need to know about implementation
details of swapchains anymore...

it also allows us to declare renderer settings with a linked list of
swapchains, which is more elegant than a fixed-size pre-allocated array
of maximum 10 elements.

+ update all examples