[backend] add post-pass explicit sync ops for images
add post pass sync for images that need it + refactor explicit sync

+ any image that is sampled just before it is then used to present on a
swapchain must have one last explicit sync so that it can be transferred
from sampled image to present source. this sync cannot be done
implicitly via renderpass/subpasses as the image is not used as an
attachment, but a regular resource.

+ move logic that adds explicit sync barriers to vulkan command buffer
into its own function so that we can use it for both pre-and post
renderpass sync barriers.