[le_window] keep track of surface lifetime
only allow one vkSurface per window. we do this by increasing an atomic
counter for surfaces in window when we create a vkSurface for this
window. This counter is decreased when the swapchain that is associated
with the surface is destroyed.

This means that the swapchain that uses the window must keep the window
alive so that it can notify it that it has destroyed the surface. We do
this by making the swapchain a part-owner of the window by increasing a
reference count on the window when the swapchain creates a surface, and
then decreasing this reference count once the surface is destroyed.