[swapchain] delay destruction until present complete
we must wait before destroying a khr_window swapchain for the current
present operation to complete on-device before we can teardown the
semaphores and images that are associated with this swapchain.

we can do that using VK_EXT_swapchain_maintenance1 which allows us to
signal a fence on completion of swapchain present.

once all these fences are signalled we know that there are no more
pending device operations and that it is safe to continue destroying the
swapchain object.