[backend] refactor
+ remove usage of std::shared_ptr as shared_ptr is possibly implemented
using vptr, which prevents hot-reloading cleanly

we should not use shared_ptr in a potentially hot-reloaded module, as
the deleter() mechanism of shared_ptr is sometimes implemented using
dynamic dispatch. This means that each shared_ptr has potentially a vptr
to a table of dynamic deleter functions - which may go stale once the
module has been reloaded. Note that this only affects a hot-reloading
scenario.