[backend] fix an issue getting default swapchain
as swapchains are stored using unordered_map, we can't just assume that
the begin() iterator points at the oldest element. We must, if there are
more than 1 swapchains, find the one with the lowest key by iterating
over all elements. The alternative would be to use a std::map (which is
ordered) to store swapchains.