[le_debug_print_text] refactor
+ remove dependency on glm as it wasn't pulling it's own weight.
+ embrace statefulness - we don't want to set cursor when printing; just
print.

We need to think about thread-safety later, if this becomes an issue;
One solution would be to calculate cursor-position and extents
atomically, and reserve these for each print instruction.

But: if we have a method gets the current cursor position, we will never
be able to guarantee that it does not get stale by the time that that
information is used...