[backend|pipeline] fix issue with set aliasing
fix an issue with descriptorset shadowing

Two sets with the same structure but with differently named bindings
would return the same (the first stored) bindings info from cache.

while it's fine to cache descriptorsets based on their structure as this
allows re-use of the vulkan objects associated with these
descriptorsets, it means that we cannot build a bindings table for each
set as only the first set would have been stored. We avoid this by
including the name of a binding in the hash for its binding info.