[backend] fix semaphore re-allocation when image not acquired
if a swapchain cannot acquire an image, we must re-create the swapchain;
this means that we must re-allocate its semaphores, too. we cannot,
however, immediately delete the old semaphores as they might still be in
use. we fix this by creating new semaphores, and swap the old semaphores
into swapchain state fields with prefix 'old_'.

once the swapchain has made a rotation (or on teardown) any semaphores
held in 'old_' fields are destroyed, which means that we have at least
one full swapchain rotation (and numSwapchainImages frameFences) between
the failed acquire and the deferred deletion of the swapchain
semaphores, which means that we can guarantee that the 'old_' semaphores
are not in-use anymore.