[resource_manager] error if width/height mismatch
in case the image has explicit extent set, and the extent does not match
the extent of the image as it appears in the file, the resource loader
will error.

we must error because the memory allocated for loading the image and the
memory required for loading the image don't match.