[backend] automatically infer image layout from binding reflection
if the binding is a sampledImage or a combinedImageSampler, layout must
be shaderReadOnlyOptimal; if there is a hint of write-access, the layout
must be General. we now store the layout with the binding, so that we don't
have to infer the layout from the bind command out of the command stream
(which would be incorrect anyway).