[renderer] refactor is-contributing logic for passes
instead of re-creating the rendergraph, update renderpasses for the
rendergraph that is being built in-place once we have resolved all the
nodes. This allows us to keep node indices in sync with pass indices,
and it also allows us to display any renderpasses that have been de-
activated for debug.