[renderer|backend] refactor of encoder
Refactor so that we can use renderpasses more flexibly, and we don't
have to worry about the lifetime of any encoders associated with the
renderpass.

We're moving encoder to purely be an adapter over command streams
which are owned by the backend.

this allows us to remove the encoder from a renderpass. ownership of
the ein a more flexible way, ncoder is with the rendergraph (but an
encoder's lifetime is limited to the duration of when we're recording
commands).

Once commands are recorded, the backend (via prefix sum) can associate
a command stream object from its current frame with each renderpass.
Since only renderpasses which contribute to the final result are
executed, we only need one command stream for each contributing
renderpass. We also don't need a command stream for any renderpasses
which don't have an exec callback. (this can be the case for
clear-only renderpasses)