[renderer] make setup callbacks one-shot
this allows us to call setup() on a rendergraph repeatedly without it
everytime calling the setup() callbacks and therefore potentially
changing user_data as a side-effect.

After we have called the setup() callback once for each pass, we want to
preserve whether this pass is being used, and which resources it does
use; we do not want to call `setup()` again, as it might change whatever
callback `user_data` points to as a side-effect. the data pointed to might
have changed as well, which means that calling setup() repeatedly would
result in different outcomes for the rendergraph.

we don't want this. we define the first call to `setup` for each pass to
be the canonical one.