[renderer] rendergraph: additionally, keep track of explicitly written-to-resources
Shader reads from images are implicitly writes, because we must
pessimistically assume that before we could read from the image
resource, we had to perform a layout transition (which is a write op).

For a rendergraph visualization, it can be more important to see where
the *contents* of an image were updated. If we additionally remember any
explicit writes, we can distinguish between implicit and explicit writes
to a recource and can better trace its history through the rendergraph.