[rendergraph_visualizer] allow recording and playback of rendergraphs
- useful if rendergraphs are dynamic and we want to be able to isolate
a rendergraph in time
- additionally reduce the api for visualizing a renderpass, don't feed
in the renderpass that needs to be visualized separately, because this
could lead to confusion over who triggers the setup() callbacks.

really, we expect the pass to be visualized to be resolved already (all
setup callbacks have been called) -- in case we are visualizing the
current rendergraph, we must resolve first before drawing, and we
must make sure not to resolve again, because calling setup() callbacks
might have changed internal state as a side-effect (for example with
ping-pong or one-shot upload resources).