[rendergraph_visualizer]  pass window extent to rendergraph initializer
rendergraph visualizer needs to know the size of the window into which
it will get drawn (so that it can correctly interpret ui mouse events,
for example).

We require an initial extents on creation. If the visualizer processes
a WindowExtent ui event, then it will update its window extents auto-
matically.