[pipeline] call maintain_cache after updating shaders
this allows all shader compilers to do whatever they need to do to
maintain their cache. For the default shader compiler this is a no-op.
For the slang shader compiler this is where we destroy the session
(cache) so that we may create a new session when compiling the next
batch of shaders.