[swapchain|backend] move present_complete semaphore into swapchain
it's not possible to be sure that a semaphore that is waited
upon by vkQueuePresent has in fact been waited upon and is in an
unsignalled state because of an API design issue.

https://docs.vulkan.org/guide/latest/swapchain_semaphore_reuse.html

we can however be certain that the semaphore **for the image index that
was successfully acquired** has been signalled; by moving ownership of
the semaphore into the swapchain we can make sure that we only re-use
semaphores that can be safely re-used. This follows the technique de-
scribed in the vulkan docs above.