[le_2d] re-compile fine_area shader
spv generated via tint had a rgba8 specifier for the output of the
storage image. this would cause issues when we wanted to render to
output images that had different pixel formats.

the root issue for this is that it appears that WGSL requires storage
texels to be declared with a format
<https://www.w3.org/TR/WGSL/#storage-texel-formats>.

This means that we had to cross-compile using this chain:

[fine_area.wgsl] --tint-> [spv] -spirv-cross-> [glsl] -glslc--> [spv]

so that re can remove the annotation from the textel storage uniform...
sigh.