[renderer|backend] add resource_usage_flag to use_resource
add a flag to use_resource that allows us to signal that we would like
the renderer to create a transient image view for this resource.

i'm not very happy with this solution, as it means meddling with
renderer internals from the outside -- the real intent of this is to
have a way to declare bindless resources to the rendergraph; right now
this is only possible by declaring the parent resources to the graph.