[backend, renderer] re-implement msaa fixes
when we are using multisampling, we must create images with higher sample
counts. These images must be associated with the renderpass; they are not
part of our regular menagerie of image resources, because they are only
used transiently for rendering -- we resolve every renderpass into its
named output image(s).

i believe there is a risk of leaking these transient image resources
when you re-allocate the swapchain -- this needs some more investigation.