[renderer] implement a callback based get_debug_name for resource handle
add a convenience method to get debug names for a resource - this is not
meant to be super fast, but correct (as it is used for debug). To
correctly use this method, you must initialize it once everytime a *compilation
unit* gets setup (that's usually when you load a module, and then whenever
a module gets reloaded via hot-reloading).

Initialization works by telling the method where to find the renderer,
and also where to find the renderer's get_resource_debug_name method.