[renderer] include parent resource handle in bindless resource handle
we use the low 32 bits of a bindless resource handle to encode the
parent resource (if any). this allows us to declare bindless resources
to the rendergraph just as we would declare "normal" resources.

Ultimately, the rendergraph is only interested in parent resources,
as these are the resources that synchronisation depends on; by
including a reference to the parent resource in a bindless handle we
allow the rendergraph to know the parent resource for a bindless handle
directly.