[le_2d] patch vello shaders so that we can use arbitrary output image formats
since wgsl only allows pre-defined format outputs to storage images, the
generated spirv does hardcode rgba8 output for the 2d rasterizer.

we want to keep the image output flexible, which is possible when using
` OpCapability StorageImageWriteWithoutFormat` in SPIRV. To patch the
shaders, we ask tint to output spirv assembly for the `fine*` shaders,
which we then patch, and assemble. This recipe is contained in
`tint_to_spv.sh`.