[le_2d] dynamically query buffer sizes
use gpu readback to inspect `vello.bump` allocator buffer for hints that
bump allocation failed. if it did, adjust the size of buffers that ran
out of space and try again.

note that this is not yet working correctly; because of the lag in
reporting an issue (the frame has to come around before we can read the
readback buffer) we are always n frames behind in correcting the required
buffer sizes. n being the number of backend frames.