[le_2d] use bump allocation size history as a low pass filter
we reduce allocation sizes only if all bump allocator history frames have
a lower value than is currently allocated. by using a history buffer we
don't get into situations where we have to re-allocate all the time,
instead we stick with the larger allocation for a while until it becomes
clear that it has not been needed.